Move evaluation context into its own document submodule
