+++ veusz/document/doc.py
@@ -24,7 +24,6 @@
 from __future__ import division
 import codecs
 import os.path
-import re
 import traceback
 import datetime
 from collections import defaultdict
@@ -36,11 +35,12 @@ try:
 except ImportError:
     h5py = None
 
-from ..compat import crange, citems, cvalues, cstr, cexec, CStringIO, cexecfile
+from ..compat import citems, cvalues, cstr, CStringIO, cexecfile
 from .. import qtall as qt4
 
 from . import widgetfactory
 from . import painthelper
+from . import evaluate
 
 from .. import datasets
 from .. import utils
@@ -49,24 +49,6 @@ from .. import setting
 def _(text, disambiguation=None, context="Document"):
     """Translate text."""
     return qt4.QCoreApplication.translate(context, text, disambiguation)
-
-# python identifier
-identifier_re = re.compile(r'^[A-Za-z_][A-Za-z0-9_]*$')
-# for splitting
-identifier_split_re = re.compile(r'[A-Za-z_][A-Za-z0-9_]*')
-
-# python module
-module_re = re.compile(r'^[A-Za-z_\.]+$')
-
-# function(arg1, arg2...) for custom functions
-# not quite correct as doesn't check for commas in correct places
-function_re = re.compile(r'''
-^([A-Za-z_][A-Za-z0-9_]*)[ ]*  # identifier
-\((                            # begin args
-(?: [ ]* ,? [ ]* [A-Za-z_][A-Za-z0-9_]* )*     # named args
-(?: [ ]* ,? [ ]* \*[A-Za-z_][A-Za-z0-9_]* )?   # *args
-(?: [ ]* ,? [ ]* \*\*[A-Za-z_][A-Za-z0-9_]* )? # **kwargs
-)\)$                           # endargs''', re.VERBOSE)
 
 def getSuitableParent(widgettype, initialwidget):
     """Find the nearest relevant parent for the widgettype given."""
@@ -125,28 +107,11 @@ class Document(qt4.QObject):
         # default document locale
         self.locale = qt4.QLocale()
 
+        # evaluation context
+        self.evaluate = evaluate.Evaluate(self)
+
         self.clearHistory()
         self.wipe()
-
-        # directories to examine when importing
-        self.importpath = []
-
-        # store custom functions and constants
-        # consists of tuples of (name, type, value)
-        # type is constant or function
-        # we use this format to preserve evaluation order
-        self.customs = []
-        self.updateEvalContext()
-
-        # copy default colormaps
-        self.colormaps = utils.ColorMaps()
-
-        # copies of validated compiled expressions
-        self.exprcompiled = {}
-        self.exprfailed = set()
-        self.exprfailedchangeset = -1
-        self.evalexprcache = {}
-        self.evalexprcachechangeset = None
 
     def wipe(self):
         """Wipe out any stored data."""
@@ -173,7 +138,7 @@ class Document(qt4.QObject):
     def enableUpdates(self):
         """Reenables document updates."""
         changeset = self.suspendupdates.pop()
-        if len(self.suspendupdates) == 0 and changeset != self.changeset:
+        if not self.suspendupdates and changeset != self.changeset:
             # bump this up as some watchers might ignore this otherwise
             self.changeset += 1
             self.setModified()
@@ -203,14 +168,9 @@ class Document(qt4.QObject):
         Updates are suspended during the operation.
         """
 
-        self.suspendUpdates()
-        try:
+        with DocSuspend(self):
             retn = operation.do(self)
             self.changeset += 1
-        except:
-            self.enableUpdates()
-            raise
-        self.enableUpdates()
 
         if self.historybatch:
             # in batch mode, create an OperationMultiple for all changes
@@ -241,14 +201,9 @@ class Document(qt4.QObject):
         """Undo the previous operation."""
 
         operation = self.historyundo.pop()
-        self.suspendUpdates()
-        try:
+        with DocSuspend(self):
             operation.undo(self)
             self.changeset += 1
-        except:
-            self.enableUpdates()
-            raise
-        self.enableUpdates()
         self.historyredo.append(operation)
 
     def canUndo(self):
@@ -473,13 +428,6 @@ class Document(qt4.QObject):
         fileobj.write('# Saved at %s\n\n' %
                       datetime.datetime.utcnow().isoformat())
 
-    def saveCustomDefinitions(self, fileobj):
-        """Save custom constants and functions."""
-
-        for vals in self.customs:
-            fileobj.write('AddCustom(%s, %s, %s)\n' %
-                          tuple([repr(x) for x in vals]))
-
     def saveDatasetTags(self, fileobj):
         """Write dataset tags to output file"""
 
@@ -493,12 +441,6 @@ class Document(qt4.QObject):
         for tag, val in sorted(bytag.items()):
             fileobj.write('TagDatasets(%s, %s)\n' %
                           (repr(tag), repr(val)))
-
-    def saveCustomFile(self, fileobj):
-        """Export the custom settings to a file."""
-
-        self._writeFileHeader(fileobj, 'custom definitions')
-        self.saveCustomDefinitions(fileobj)
 
     def saveToFile(self, fileobj):
         """Save the text representing a document to a file.
@@ -521,7 +463,7 @@ class Document(qt4.QObject):
             fileobj.write('AddImportPath(%s)\n' % repr(reldirname))
 
         # add custom definitions
-        self.saveCustomDefinitions(fileobj)
+        self.evaluate.saveCustomDefinitions(fileobj)
 
         # save those datasets which are linked
         # we do this first in case the datasets are overridden below
@@ -564,7 +506,7 @@ class Document(qt4.QObject):
             textstream.write('AddImportPath(%s)\n' % repr(reldirname))
 
         # add custom definitions
-        self.saveCustomDefinitions(textstream)
+        self.evaluate.saveCustomDefinitions(textstream)
 
         # save those datasets which are linked
         # we do this first in case the datasets are overridden below
@@ -745,302 +687,6 @@ class Document(qt4.QObject):
         # return widget
         return obj
 
-    def _processSafeImports(self, module, symbols):
-        """Check what symbols are safe to import."""
-
-        # empty list
-        if not symbols:
-            return symbols
-
-        # do import anyway
-        if setting.transient_settings['unsafe_mode']:
-            return symbols
-
-        # two-pass to ask user whether they want to import symbol
-        for thepass in crange(2):
-            # remembered during session
-            a = 'import_allowed'
-            if a not in setting.transient_settings:
-                setting.transient_settings[a] = defaultdict(set)
-            allowed = setting.transient_settings[a][module]
-
-            # not allowed during session
-            a = 'import_notallowed'
-            if a not in setting.transient_settings:
-                setting.transient_settings[a] = defaultdict(set)
-            notallowed = setting.transient_settings[a][module]
-
-            # remembered in setting file
-            a = 'import_allowed'
-            if a not in setting.settingdb:
-                setting.settingdb[a] = {}
-            if module not in setting.settingdb[a]:
-                setting.settingdb[a][module] = {}
-            allowed_always = setting.settingdb[a][module]
-
-            # collect up
-            toimport = []
-            possibleimport = []
-            for symbol in symbols:
-                if symbol in allowed or symbol in allowed_always:
-                    toimport.append(symbol)
-                elif symbol not in notallowed:
-                    possibleimport.append(symbol)
-
-            # nothing to do, so leave
-            if not possibleimport:
-                break
-
-            # only ask the user the first time
-            if thepass == 0:
-                self.sigAllowedImports.emit(module, possibleimport)
-
-        return toimport
-
-    def _updateEvalContextImport(self, module, val):
-        """Add an import statement to the eval function context."""
-        if module_re.match(module):
-            # work out what is safe to import
-            symbols = identifier_split_re.findall(val)
-            toimport = self._processSafeImports(module, symbols)
-            if toimport:
-                defn = 'from %s import %s' % (module,
-                                              ', '.join(toimport))
-                try:
-                    cexec(defn, self.eval_context)
-                except Exception:
-                    self.log(_("Failed to import '%s' from "
-                               "module '%s'") % (', '.join(toimport),
-                                                 module))
-                    return
-
-            delta = set(symbols)-set(toimport)
-            if delta:
-                self.log(_("Did not import '%s' from module '%s'") %
-                         (', '.join(list(delta)), module))
-
-        else:
-            self.log( _("Invalid module name '%s'") % module )
-
-    def validateProcessColormap(self, colormap):
-        """Validate and process a colormap value.
-
-        Returns a list of B,G,R,alpha tuples or raises ValueError if a problem."""
-
-        try:
-            if len(colormap) < 2:
-                raise ValueError( _("Need at least two entries in colormap") )
-        except TypeError:
-            raise ValueError( _("Invalid type for colormap") )
-
-        out = []
-        for entry in colormap:
-            for v in entry:
-                try:
-                    v - 0
-                except TypeError:
-                    raise ValueError(
-                        _("Colormap entries should be numerical") )
-                if v < 0 or v > 255:
-                    raise ValueError(
-                        _("Colormap entries should be between 0 and 255") )
-
-            if len(entry) == 3:
-                out.append( (int(entry[2]), int(entry[1]), int(entry[0]),
-                             255) )
-            elif len(entry) == 4:
-                out.append( (int(entry[2]), int(entry[1]), int(entry[0]),
-                             int(entry[3])) )
-            else:
-                raise ValueError( _("Each colormap entry consists of R,G,B "
-                                    "and optionally alpha values") )
-
-        return tuple(out)
-
-    def _updateEvalContextColormap(self, name, val):
-        """Add a colormap entry."""
-
-        try:
-            cmap = self.validateProcessColormap(val)
-        except ValueError as e:
-            self.log( cstr(e) )
-        else:
-            self.colormaps[ cstr(name) ] = cmap
-
-    def _updateEvalContextFuncOrConst(self, ctype, name, val):
-        """Update a function or constant in eval function context."""
-
-        if ctype == 'constant':
-            if not identifier_re.match(name):
-                self.log( _("Invalid constant name '%s'") % name )
-                return
-            defn = val
-        elif ctype == 'function':
-            m = function_re.match(name)
-            if not m:
-                self.log( _("Invalid function specification '%s'") % name )
-                return
-            name = m.group(1)
-            args = m.group(2)
-            defn = 'lambda %s: %s' % (args, val)
-
-        # evaluate, but we ignore any unsafe commands or exceptions
-        comp = self.compileCheckedExpression(defn)
-        if comp is None:
-            return
-        try:
-            self.eval_context[name] = eval(comp, self.eval_context)
-        except Exception as e:
-            self.log( _("Error evaluating '%s': '%s'") %
-                      (name, cstr(e)) )
-
-    def compileCheckedExpression(self, expr, origexpr=None, log=True):
-        """Compile expression and check for errors.
-
-        origexpr is an expression to show in error messages. This is
-        used if replacements have been done, etc.
-        """
-
-        try:
-            return self.exprcompiled[expr]
-        except KeyError:
-            pass
-
-        # track failed compilations, so we only print them once
-        if self.exprfailedchangeset != self.changeset:
-            self.exprfailedchangeset = self.changeset
-            self.exprfailed.clear()
-        elif expr in self.exprfailed:
-            return None
-
-        if origexpr is None:
-            origexpr = expr
-
-        try:
-            checked = utils.compileChecked(
-                expr,
-                ignoresecurity=setting.transient_settings['unsafe_mode'])
-        except utils.SafeEvalException as e:
-            if log:
-                self.log(
-                    _("Unsafe expression '%s': %s") % (origexpr, cstr(e)))
-            self.exprfailed.add(expr)
-            return None
-        except Exception as e:
-            if log:
-                self.log(
-                    _("Error in expression '%s': %s") % (origexpr, cstr(e)))
-            return None
-        else:
-            self.exprcompiled[expr] = checked
-            return checked
-
-    @staticmethod
-    def _evalformatdate(fmt=None):
-        """DATE() eval: return date with optional format."""
-        d = datetime.date.today()
-        return d.isoformat() if fmt is None else d.strftime(fmt)
-
-    @staticmethod
-    def _evalformattime(fmt=None):
-        """TIME() eval: return time with optional format."""
-        t = datetime.datetime.now()
-        return t.isoformat() if fmt is None else t.strftime(fmt)
-
-    def _evaldata(self, name, part='data'):
-        """DATA(name, [part]) eval: return dataset as array."""
-        if part not in ('data', 'perr', 'serr', 'nerr'):
-            raise RuntimeError("Invalid dataset part '%s'" % part)
-        if name not in self.data:
-            raise RuntimeError("Dataset '%s' does not exist" % name)
-        data = getattr(self.data[name], part)
-        if isinstance(data, N.ndarray):
-            return N.array(data)
-        elif isinstance(data, list):
-            return list(data)
-        return data
-
-    def _evalfilename(self):
-        """FILENAME() eval: returns filename."""
-        return utils.latexEscape(self.filename)
-
-    def _evalbasename(self):
-        """BASENAME() eval: returns base filename."""
-        return utils.latexEscape(os.path.basename(self.filename))
-
-    def _evalsetting(self, path):
-        """SETTING() eval: return setting given full path."""
-        return self.resolveFullSettingPath(path).get()
-
-    def updateEvalContext(self):
-        """To be called after custom constants or functions are changed.
-        This sets up a safe environment where things can be evaluated
-        """
-
-        self.eval_context = c = {}
-
-        # add numpy things
-        # we try to avoid various bits and pieces for safety
-        for name, val in citems(N.__dict__):
-            if ( (callable(val) or type(val)==float) and
-                 name not in __builtins__ and
-                 name[:1] != '_' and name[-1:] != '_' ):
-                c[name] = val
-
-        # safe functions
-        c['os_path_join'] = os.path.join
-        c['os_path_dirname'] = os.path.dirname
-        c['veusz_markercodes'] = tuple(utils.MarkerCodes)
-
-        # helpful functions for expansion
-        c['ENVIRON'] = dict(os.environ)
-        c['DATE'] = self._evalformatdate
-        c['TIME'] = self._evalformattime
-        c['DATA'] = self._evaldata
-        c['FILENAME'] = self._evalfilename
-        c['BASENAME'] = self._evalbasename
-        c['ESCAPE'] = utils.latexEscape
-        c['SETTING'] = self._evalsetting
-
-        # custom definitions
-        for ctype, name, val in self.customs:
-            name = name.strip()
-            if ctype == 'constant' or ctype == 'function':
-                self._updateEvalContextFuncOrConst(ctype, name, val.strip())
-            elif ctype == 'import':
-                self._updateEvalContextImport(name, val)
-            elif ctype == 'colormap':
-                self._updateEvalContextColormap(name, val)
-            else:
-                raise ValueError('Invalid custom type')
-
-    def customDict(self):
-        """Return a dictionary mapping custom names to (idx, type, value)."""
-        retn = {}
-        for i, (ctype, name, val) in enumerate(self.customs):
-            retn[name] = (i, ctype, val)
-        return retn
-
-    def evalDatasetExpression(self, expr, part='data', datatype='numeric',
-                              dimensions=1):
-        """Return dataset after evaluating a dataset expression.
-        part is 'data', 'serr', 'perr' or 'nerr' - these are the
-        dataset parts which are evaluated by the expression
-
-        None is returned on error
-        """
-
-        key = (expr, part, datatype, dimensions)
-        if self.evalexprcachechangeset != self.changeset:
-            self.evalexprcachechangeset = self.changeset
-            self.evalexprcache.clear()
-        elif key in self.evalexprcache:
-            return self.evalexprcache[key]
-
-        self.evalexprcache[key] = ds = datasets.evalDatasetExpression(
-            self, expr, part=part, datatype=datatype, dimensions=dimensions)
-        return ds
-
     def valsToDataset(self, vals, datatype, dimensions):
         """Return a dataset given a numpy array of values."""
 
@@ -1098,17 +744,6 @@ class Document(qt4.QObject):
                 self.walkNodes(tocall, root=s, nodetypes=nodetypes,
                                _path = _path + '/' + s.name)
         # elif root.nodetype == 'setting': pass
-
-    def getColormap(self, name, invert):
-        """Get colormap with name given (returning grey if does not exist)."""
-        cmap = self.colormaps.get(name, self.colormaps['grey'])
-        if invert:
-            if cmap[0][0] >= 0:
-                return cmap[::-1]
-            else:
-                # ignore marker at beginning for stepped maps
-                return tuple([cmap[0]] + list(cmap[-1:0:-1]))
-        return cmap
 
     def formatValsWithDatatypeToText(self, vals, displaydatatype):
         """Given a set of values, datatype, return a list of strings
