+++ veusz/setting/controls.py
@@ -1836,7 +1836,7 @@ class Colormap(Choice):
     size = (32, 12)
 
     def __init__(self, setn, document, parent):
-        names = sorted(document.colormaps)
+        names = sorted(document.evaluate.colormaps)
 
         icons = Colormap._generateIcons(document, names)
         Choice.__init__(self, setn, True,
@@ -1857,7 +1857,7 @@ class Colormap(Choice):
 
         # iterate over colour maps
         for name in names:
-            val = document.colormaps.get(name, None)
+            val = document.evaluate.colormaps.get(name, None)
             if name in kls._icons:
                 icon = kls._icons[name]
             else:
